% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dnacheck.R
\name{dnacheck}
\alias{dnacheck}
\title{Check if the DNA sequence are in the 4 default types}
\usage{
dnacheck(x)
}
\arguments{
\item{x}{A character vector, as the input DNA sequence.}
}
\value{
Logical. \code{TRUE} if all of the DNA types of the sequence
        are within the 4 default types.

The result character vector
}
\description{
Check if the DNA sequence are in the 4 default types
}
\details{
This function checks if the DNA sequence types are in the 4.
}
\examples{
x = 'GACTGAACTGCACTTTGGTTTCATATTATTTGCTC'
dnacheck(x) # TRUE
dnacheck(paste(x, 'Z', sep = '')) # FALSE
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>
}
\keyword{check}

